security/oidc: Move request_auth from utils to security
