+++ src/v/security/request_auth.h
